rename: 목업데이터 도메인 기준으로 재분류 feature/TODAK-197/update-mockup
