+++ components/layout/BottomNavigation.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { useRouter } from "next/navigation";
-import path from "@/feature/routes";
+import path from "@/routes";
 import { HomeSVG, PencilSVG, UserSVG } from "../svg";
 import { useHideOnScroll } from "@/utils/hooks";
