+++ domain/auth/mocks/auth.ts
