+++ domain/diary/mocks/diary.ts
