--- feature/mocks/config/handlers.ts
@@ -1,11 +1,0 @@
-import { authMockups } from "../domain/auth";
-import { feedMockups } from "../domain/feed";
-import { memberMockups } from "../domain/member";
-import { diaryMockups } from "../domain/diary";
-
-export const handlers = [
-  ...authMockups,
-  ...feedMockups,
-  ...memberMockups,
-  ...diaryMockups,
-];
