+++ mocks/config/browser.ts
